Change cfg_high/cfg_low to float; add Enter key focus advance
- cfg_high and cfg_low are now FLOAT throughout (Python types, JS payloads,
  step=0.1 on inputs, README table)
- Enter key on any text/number input in the edit form moves focus to the
  next field (skips file/hidden inputs; textareas keep default newline)

+++ nodes/workflow_config_wan22.py
@@ -78,7 +78,7 @@ class WorkflowConfigWan22:
         "IMAGE",
         "INT", "INT", "INT", "INT",
         "STRING", "STRING", "STRING",
-        "INT", "INT", "INT",
+        "FLOAT", "FLOAT", "INT",
         "FLOAT",
     )
     RETURN_NAMES = (
@@ -121,8 +121,8 @@ class WorkflowConfigWan22:
             str(entry.get("master_prompt",   "")),
             str(entry.get("positive_prompt", "")),
             str(entry.get("negative_prompt", "")),
-            int(entry.get("cfg_high",     0)),
-            int(entry.get("cfg_low",      0)),
+            float(entry.get("cfg_high",  0.0)),
+            float(entry.get("cfg_low",   0.0)),
             int(entry.get("total_frames", 0)),
             float(entry.get("fps", 0.0)),
         )
